+++ Core/App/ActivitySchedulerApp.cs
@@ -1,16 +1,24 @@
-﻿using System;
+﻿using ActivityScheduler.Shared;
+using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.ServiceProcess;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
+using System.Windows.Interop;
 
 namespace ActivityScheduler.Core.Appilcation
 {
     public class ActivitySchedulerApp
     {
 
-        public ActivitySchedulerApp() { }
+        public ActivitySchedulerApp() 
+        {
+            
+        }
 
         public String? BaseDirectory { get => System.IO.Path.GetDirectoryName(System.Reflection.Assembly.GetExecutingAssembly().Location); }
 
@@ -52,6 +60,19 @@ namespace ActivityScheduler.Core.Appilcation
                 return directory;
             }
         }
+        
+        public String? ServiceExeFileFullPath
+        {
+            get
+            {
+                var directory = "C:\\Develop\\ActivityScheduler\\ActivityScheduler.WorkerService.TopShelf\\bin\\Debug\\net6.0\\ActivityScheduler.WorkerService.TopShelf.exe";
+                return directory;
+            }
+        }
 
+
+        public String? WinServiceName => "A01";
+        public String? WinServiceDiaplayName => "A01_Display_Name";
+        public String? WinServiceDescription => "A01_Description";
     }
 }