+++ Core/Settings/Settings.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using ActivityScheduler.Core.Appilcation;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -16,10 +17,14 @@ namespace ActivityScheduler.Core.Settings
 {
     public partial class Settings : Window
     {
-        SettingsManager _settingsManager;
-        public Settings(SettingsManager settingsManager)
+        private SettingsManager _settingsManager;
+        private ActivitySchedulerApp _app;
+        private WorkerServiceManager _workerMgr;
+        public Settings(SettingsManager settingsManager, ActivitySchedulerApp app, WorkerServiceManager workerMgr)
         {
             _settingsManager = settingsManager;
+            _app = app;
+            _workerMgr = workerMgr;
             InitializeComponent();
         }
 
@@ -48,6 +53,36 @@ namespace ActivityScheduler.Core.Settings
             Setting1Tb.Text = settings.Setting01;
             
             Setting2Chbx.IsChecked = settings.Setting02;
+        }
+
+        private void Button_Click(object sender, RoutedEventArgs e)
+        {
+
+        }
+
+        private void GetStateBtn_Click(object sender, RoutedEventArgs e)
+        {
+            StateLbl.Text = _workerMgr.GetServiceState().ToString();
+        }
+
+        private void InstallBtn_Click(object sender, RoutedEventArgs e)
+        {
+            _workerMgr.InstallService();
+        }
+
+        private void UninstallBtn_Click(object sender, RoutedEventArgs e)
+        {
+            _workerMgr.UninstallService();
+        }
+
+        private void RunBtn_Click(object sender, RoutedEventArgs e)
+        {
+            _workerMgr.StartService();
+        }
+
+        private void StopBtn_Click(object sender, RoutedEventArgs e)
+        {
+            _workerMgr.StopService();
         }
     }
 }