+++ DataAccess/IAsyncRepositoryT.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
+using ActivityScheduler.Shared;
 
 namespace ActivityScheduler.DataAccess
 {