+++ Releases.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+- #11: Create simple windows-service app that will start main
+- #12: The only one instance of application should be opened
+- #15: App doesnt show visual output when I start it from windows service (bug)
+- #16: Organize communication between worker service and WPF app via named pipes (task)
+
 ## 1.0.0
 
 - #10: Create WPF app basis with logger and settings on EF-SQLITE